# refactor(pkg): availability caching