* Move global state to the solver context
* Cache the result

Signed-off-by: Rudi Grinberg <me@rgrinberg.com>

<!-- ps-id: d0fc8e2a-9abe-47de-b674-41708f672b78 -->