@Alizter 2023-11-22
### Main:
```
Benchmark 1: ~/dune/_build/install/default/bin/dune pkg lock --skip-update
  Time (mean ± σ):      7.976 s ±  0.381 s    [User: 7.142 s, System: 0.809 s]
  Range (min … max):    7.435 s …  8.414 s    10 runs
```

### 	PR:
```
Benchmark 1: ~/dune/_build/install/default/bin/dune pkg lock --skip-update
  Time (mean ± σ):      7.561 s ±  0.094 s    [User: 6.799 s, System: 0.732 s]
  Range (min … max):    7.428 s …  7.766 s    10 runs
```

Looks like it cut down a lot of uncertainty due to reduced IO. :+1: